// src/@types/types.d.ts
type Maybe<T> = T | null;
type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigInt: any;
  Coordinates: any;
  ISO8601Date: any;
  ISO8601DateTime: any;
  URL: any;
};

type Address = {
  __typename?: "Address";
  cityCode: Scalars["String"];
  cityName: Scalars["String"];
  departmentCode?: Maybe<Scalars["String"]>;
  departmentName?: Maybe<Scalars["String"]>;
  geometry?: Maybe<GeoJson>;
  label: Scalars["String"];
  postalCode: Scalars["String"];
  regionCode?: Maybe<Scalars["String"]>;
  regionName?: Maybe<Scalars["String"]>;
  streetAddress?: Maybe<Scalars["String"]>;
  streetName?: Maybe<Scalars["String"]>;
  streetNumber?: Maybe<Scalars["String"]>;
  type: AddressType;
};
// ...
type Avis = {
  __typename?: "Avis";
  attachment?: Maybe<File>;
  claimant?: Maybe<Profile>;
  dateQuestion: Scalars["ISO8601DateTime"];
  dateReponse?: Maybe<Scalars["ISO8601DateTime"]>;
  expert?: Maybe<Profile>;
  id: Scalars["ID"];
  /** @deprecated Utilisez le champ `claimant` à la place. */
  instructeur: Profile;
  question: Scalars["String"];
  reponse?: Maybe<Scalars["String"]>;
};

type CarteChamp = Champ & {
  __typename?: "CarteChamp";
  geoAreas: Array<GeoArea>;
  id: Scalars["ID"];
  label: Scalars["String"];
  stringValue?: Maybe<Scalars["String"]>;
};

type Champ = {
  id: Scalars["ID"];
  label: Scalars["String"];
  stringValue?: Maybe<Scalars["String"]>;
};

type ChampDescriptor = {
  __typename?: "ChampDescriptor";
  champDescriptors?: Maybe<Array<ChampDescriptor>>;
  description?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  label: Scalars["String"];
  options?: Maybe<Array<Scalars["String"]>>;
  required: Scalars["Boolean"];
  type: TypeDeChamp;
};

type CheckboxChamp = Champ & {
  __typename?: "CheckboxChamp";
  id: Scalars["ID"];
  label: Scalars["String"];
  stringValue?: Maybe<Scalars["String"]>;
  value: Scalars["Boolean"];
};

enum Civilite {
  M = "M",
  Mme = "Mme",
}
// ...
type Demandeur = {
  id: Scalars["ID"];
};
// ...
type DemarcheDescriptor = {
  __typename?: "DemarcheDescriptor";
  cadreJuridique?: Maybe<Scalars["String"]>;
  dateCreation: Scalars["ISO8601DateTime"];
  dateDepublication?: Maybe<Scalars["ISO8601DateTime"]>;
  dateDerniereModification: Scalars["ISO8601DateTime"];
  dateFermeture?: Maybe<Scalars["ISO8601DateTime"]>;
  datePublication?: Maybe<Scalars["ISO8601DateTime"]>;
  declarative?: Maybe<DossierDeclarativeState>;
  deliberation?: Maybe<Scalars["String"]>;
  description: Scalars["String"];
  id: Scalars["ID"];
  number: Scalars["Int"];
  revision: Revision;
  service?: Maybe<Service>;
  state: DemarcheState;
  title: Scalars["String"];
};

enum DemarcheState {
  Brouillon = "brouillon",
  Publiee = "publiee",
  Close = "close",
  Depubliee = "depubliee",
}

type Departement = {
  __typename?: "Departement";
  code: Scalars["String"];
  name: Scalars["String"];
};

type DirectUpload = {
  __typename?: "DirectUpload";
  blobId: Scalars["ID"];
  headers: Scalars["String"];
  signedBlobId: Scalars["ID"];
  url: Scalars["String"];
};

type Dossier = {
  __typename?: "Dossier";
  annotations: Array<Champ>;
  archived: Scalars["Boolean"];
  attestation?: Maybe<File>;
  avis: Array<Avis>;
  champs: Array<Champ>;
  dateDepot: Scalars["ISO8601DateTime"];
  dateDerniereModification: Scalars["ISO8601DateTime"];
  dateExpiration?: Maybe<Scalars["ISO8601DateTime"]>;
  datePassageEnConstruction: Scalars["ISO8601DateTime"];
  datePassageEnInstruction?: Maybe<Scalars["ISO8601DateTime"]>;
  dateSuppressionParAdministration?: Maybe<Scalars["ISO8601DateTime"]>;
  dateSuppressionParUsager?: Maybe<Scalars["ISO8601DateTime"]>;
  dateTraitement?: Maybe<Scalars["ISO8601DateTime"]>;
  demandeur: Demandeur;
  demarche: DemarcheDescriptor;
  geojson?: Maybe<File>;
  groupeInstructeur: GroupeInstructeur;
  id: Scalars["ID"];
  instructeurs: Array<Profile>;
  messages: Array<Message>;
  motivation?: Maybe<Scalars["String"]>;
  motivationAttachment?: Maybe<File>;
  number: Scalars["Int"];
  pdf?: Maybe<File>;
  /** @deprecated Utilisez le champ `demarche.revision` à la place. */
  revision: Revision;
  state: DossierState;
  traitements: Array<Traitement>;
  usager: Profile;
};
// ...
enum DossierDeclarativeState {
  EnInstruction = "en_instruction",
  Accepte = "accepte",
}
// ...
enum DossierState {
  EnConstruction = "en_construction",
  EnInstruction = "en_instruction",
  Accepte = "accepte",
  Refuse = "refuse",
  SansSuite = "sans_suite",
}
// ...
type File = {
  __typename?: "File";
  /** @deprecated Utilisez le champ `byteSizeBigInt` à la place. */
  byteSize: Scalars["Int"];
  byteSizeBigInt: Scalars["BigInt"];
  checksum: Scalars["String"];
  contentType: Scalars["String"];
  filename: Scalars["String"];
  url: Scalars["URL"];
};

type GeoArea = {
  description?: Maybe<Scalars["String"]>;
  geometry: GeoJson;
  id: Scalars["ID"];
  source: GeoAreaSource;
};

enum GeoAreaSource {
  Cadastre = "cadastre",
  SelectionUtilisateur = "selection_utilisateur",
}

type GeoJson = {
  __typename?: "GeoJSON";
  coordinates: Scalars["Coordinates"];
  type: Scalars["String"];
};

type GroupeInstructeur = {
  __typename?: "GroupeInstructeur";
  id: Scalars["ID"];
  instructeurs: Array<Profile>;
  label: Scalars["String"];
  number: Scalars["Int"];
};
// ...
type Message = {
  __typename?: "Message";
  attachment?: Maybe<File>;
  body: Scalars["String"];
  createdAt: Scalars["ISO8601DateTime"];
  email: Scalars["String"];
  id: Scalars["ID"];
};
// ...
type Profile = {
  __typename?: "Profile";
  email: Scalars["String"];
  id: Scalars["ID"];
};
// ...
type Revision = {
  __typename?: "Revision";
  annotationDescriptors: Array<ChampDescriptor>;
  champDescriptors: Array<ChampDescriptor>;
  dateCreation: Scalars["ISO8601DateTime"];
  datePublication?: Maybe<Scalars["ISO8601DateTime"]>;
  id: Scalars["ID"];
};

type SelectionUtilisateur = GeoArea & {
  __typename?: "SelectionUtilisateur";
  description?: Maybe<Scalars["String"]>;
  geometry: GeoJson;
  id: Scalars["ID"];
  source: GeoAreaSource;
};

type Service = {
  __typename?: "Service";
  id: Scalars["ID"];
  nom: Scalars["String"];
  organisme: Scalars["String"];
  siret?: Maybe<Scalars["String"]>;
  typeOrganisme: TypeOrganisme;
};
// ...
type Traitement = {
  __typename?: "Traitement";
  dateTraitement: Scalars["ISO8601DateTime"];
  emailAgentTraitant?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  motivation?: Maybe<Scalars["String"]>;
  state: DossierState;
};

enum TypeDeChamp {
  Text = "text",
  Textarea = "textarea",
  Date = "date",
  Datetime = "datetime",
  Number = "number",
  DecimalNumber = "decimal_number",
  IntegerNumber = "integer_number",
  Checkbox = "checkbox",
  Civilite = "civilite",
  Email = "email",
  Phone = "phone",
  Address = "address",
  YesNo = "yes_no",
  DropDownList = "drop_down_list",
  MultipleDropDownList = "multiple_drop_down_list",
  LinkedDropDownList = "linked_drop_down_list",
  Pays = "pays",
  Regions = "regions",
  Departements = "departements",
  Communes = "communes",
  Engagement = "engagement",
  HeaderSection = "header_section",
  Explication = "explication",
  DossierLink = "dossier_link",
  PieceJustificative = "piece_justificative",
  Siret = "siret",
  Carte = "carte",
  Repetition = "repetition",
  TitreIdentite = "titre_identite",
  Iban = "iban",
  AnnuaireEducation = "annuaire_education",
  Cnaf = "cnaf",
  Dgfip = "dgfip",
  PoleEmploi = "pole_emploi",
  Mesri = "mesri",
}